Implement the parallel CREATE INDEX step.
At the moment the index jobs limit is not taken into account, we just fork
as many sub-processes as we have to create indexes for each table. We also
create constraints attached to indexes from their definitions, once the
index is ready. This allows even primary key indexes to be built in parallel
to the other indexes on a same table.