Implement --index-jobs option.
This allows controling the concurrency of CREATE INDEX operations separately
from the concurrency of the COPY operations. Also, the index operations are
limited globally, not per-table.

After all it makes sense for the users to setup how many CPUs they want to
be active creating indexes at the same time globally.