Implement some tests, using pagila.
This is a first step towards automated testing. Two Postgres instances are
deployed, the pagila database is deployed in the source instance and then
copied over to the target instance using pgcopydb.

The next step consists of using a Postgres schema diff tool to check that
the source and target schemas are indeed the same.