Implement --restart, --resume --not-consistent.
This allows to resume from a previous run that would have been interrupted,
or would have failed halfway through. To be able to continue from a previous
run, the user needs to accept a non-consistent operation, where another
snapshot is used the second time.

Internally, the work that has been done already is tracked via lock files
and done files in the working directory.