Use the source database snapshot in the schema pg_dump steps.
In the previous coding we would first call into pg_dump to grab the pre-data
and post-data schemas to file, and then only start a transaction and export
its snapshot for re-use in the table data COPY processes.

With this patch we now take the snapshot first and then use the exported
snapshot in the pg_dump calls and also in the rest of the operations,
ensuring consistency in between even those steps.