First batch of work to implement support for logical decoding.
The idea is to enable Change data capture while the initial copy of the
source database is happening, in a way that we can then later catch-up with
the changes and reduce the maintenance window for the migration.

The trouble with using Logical Replication directly is that:

  - logical replication has no support for DDLs, so the schema parts needs a
    separate handling anyways

  - during the initial copy of the data the logical replication subscriber
    does not consume from the replication slot, meaning that a high traffic
	source system might end-up filling up the WAL disk entirely when the
	initial copy takes too long

  - using logical replication requires creating new SQL objects in both the
    source and the target database, and the subscription on the target
	database requires superuser access.

All in all there is a strong case to be made for writing a client-side
logical decoding client that replays the changes from the source system into
the target system.