Fix pgcopydb copy constraints.
Instead of using the `pgcopydb copy data` approach focused on tables, and
then for each tables the indexes, and then when all the indexes of a given
table are finished creating the constraints; simplify and re-use the index
only code path from `pgcopydb copy indexes` taking care of constraints.

This means indexJobs processes are created and share the list of constraints
to build with the usual locking approach.