Allow using --resume with --snapshot without --not-consistent. (#13)
* Allow using --resume with --snapshot without --not-consistent.

When the snapshot has been kept around in between two runs of pgcopydb we
can still ensure consistency and use --resume and --snapshot.

* Improve --resume --snapshot ... processing.

Make the processing more robust when using `pgcopydb copy ...` commands and
refactor some code to share more of the tooling. Specifically, all the
commands now share the same getopt structure and support the same
environment variables.

* Add proper support for --snapshot for pgcopydb copy ... subcommands.

* Fix pgcopydb copy constraints.

Instead of using the `pgcopydb copy data` approach focused on tables, and
then for each tables the indexes, and then when all the indexes of a given
table are finished creating the constraints; simplify and re-use the index
only code path from `pgcopydb copy indexes` taking care of constraints.

This means indexJobs processes are created and share the list of constraints
to build with the usual locking approach.