Implement setting GUC values to our src/dst Postgres connections.
At the minimum we want to use the same client_encoding on both connections,
so that we know that we are dealing with the same data without having to
actually look into it.

Then, for bulk-loading and index creation, some settings are easy to set on
the target connection to get a better perf profile (e.g.
maintenance_work_mem).