Implement support for Large Objects (#20)
* pgcopydb copy blobs

New command to support copying large objects from the source to the target
database instances, in a streaming fashion: we avoid spilling the large
object data to disk altogether.

* Integrate an auxiliary process in the main table-data part.

Now the large objects are automatically copied over from the source to the
target database. The concurrency is limited to 1 by the way the code is
written at the moment.