Bug fix: recognize index supporting constraints.
The "constraint" boolean arg is about our intend to either create the index
or the constraint (that maybe attached to the index). Our decision making
about the concurrent building of indexes should apply to building an index
that is baking a constraint that's neither PK or UNIQUE, we can't re-use our
"constraint" boolean for that decision making.