Implement Unit/Regression testing suite. (#34)
* Implement Unit/Regression testing suite.

At the moment it's quite a simple test suite, and only covers the exclusion
constraint issue #22. As more specific issues are opened, we can add to this
regression test suite.

* Add SQL based checking.

To ensure that the test is passing, it is now possible to run SQL tests and
compare the result from the registered expected output.