Fix support for inherited tables.
A good optimization when using COPY is to use TRUNCATE in the same
transaction, so as to be able to then use COPY FREEZE and avoid some VACUUM
activities after the bulk loading of data.

Trouble is that with inherited tables, the TRUNCATE command would also
remove all the data that was (maybe) just copied to the inheritance tree
nodes.

To fix, just use the TRUNCATE ONLY variant of the command.

This commit also adds full regression testing for this issue, thanks to the
test case provided in #33.