Improve filtering of pg_restore list entries without an OID.
Some pg_restore --list entries such as INDEX ATTACH miss both the catalog
and the object oid, and then we need to match them by their pg_restore list
name, which is a compound of the schema name, object name, and owner name.

Adding to that, the way pg_restore builds that compound name is with using a
single space as a separator, and replacing \n and \r characters with a
single space. This makes the pg_restore list output unfriendly to machine
parsing, and so instead we generate the list name from the Postgres catalogs
in our catalog queries.

Then we can use an hash table on the OIDs and another one on the compound
names and find out if an INDEX ATTACH catalog entry refers to an index that
has been filtered-out by filtering rules, and comment it out then.