Edit the default GIT-VERSION-GEN version.
We still have a problem with how the version string is computed in the
Docker image, where there is no .git environment and then no tags. The
script from the git git repository uses an extra "version" file for
releases, and that's not what we're doing at the moment.

We might want to produce the version file on the local machine and then send
it in the Docker environment, but at the moment, let's just keep it simple.