Fix filtering of the pre-data section of the dump. (#44)
* Implement filtering the pre-data section.

When installing the schema on the target database, we want to already filter
out the SQL objects that are filtered-out in the filtering setup. The
previous coding forgot about that parts.

See #43.

* Implement support for exclude-schemas.

The parsing was implemented already, but the filtering itself was shy of a
brick load. We need to both filter tables that belong to the schema (we did)
and also the schema object itself (we didn't).

* Add missed SQL objects types to parse from pg_restore archive list.

* Add ACL as a pg_restore archive catalog description entry.