Skip long-running transactions when --not-consistent is used.
To prevent limitations around lock management when running pgcopydb on a
source database with lots of objects, we now benefit from --not-consistent
as an opportunity to reduce long-running transactions.

It was already the case that --not-consistent wouldn't require exporting or
setting a snapshot, but the opportunity was not yet implemented fully.