Set tcp_keepalives_idle to 60s on the source database.
Because we might keep open a long-running transaction on the source
database, the one that keeps a snapshot open for the whole duration of the
migration, then it's probably best to set the tcp_keepalives_idle to
something short.

This should help avoid some problems with firewalls shutting down long
running (idle) Postgres connections or transactions.