Implement pgcopydb stream apply.
Applying the SQL is as easy as reading the prepared SQL file one line at a
time and then sending that SQL query to the target system, though it also
requires using the Postgres Replication Progress Tracking facility to skip
previously applied transactions.

This patch does not implement cleaning the replication origin tracking on
the target system at the moment.