Implement support for copying roles. (#64)
* Implement support for copying roles.

Postgres roles are global objects, which means they do not belong to any
specific database in an instance. To be able to dump and restore the roles,
we need to use pg_dumpall --roles-only.

* Manually send SQL commands from the pg_dumpall --roles-only script.

This allows to implement "create role if not exists" our own way, accepting
roles that already exists as they are.

* Have tests/pagila use a non-postgres role.

We now create a "pagila" role and make it owner of the database and its
objects, and then run the pgcopydb commands using that new non-superuser
role.

The creation of the new role on the target system is made by calling into
pgcopydb dump roles and pgcopydb restore roles.

* Skip ALTER ROLE commands too when skipping CREATE ROLE.

* Imnplement pgcopydb copy roles.

* Implement pgcopydb clone --roles.

This includes a STEP 0 that copies the roles from the source instance to the
target instance.