WIP: Implement a sentinel.
The sentinel is a table on the source database with 3 columns to share the
LSN start position for the logical decoding receiver, the LSN end position
where to stop receiving and applying logical decoding changes, and finally
the apply boolean which can be set to true when the catching-up process is
allowed to start processing pre-fetched events.

The idea is to set the sentinel apply boolean as soon as the base copy
finishes, and to allow pgcopydb users and test suite to set and endpos after
the start of the follow processes.