Review process organisation for pgcopydb clone --follow.
We need to separate the clone process tree (with its calls to waipid()
blocking until all the sub-processes are done) and the follow process tree
that creates two extra sub-processes (prefetch and catchup).

As a result `pgcopydb clone` process tree is impacted too, and the main
process only exports a snapshot now. The rest of the top-level processing is
actually implemented in the "main" (or "clone") sub-process, even when a
"follow" sub-process is not started.