When pgcopydb clone --follow is used, setup before forking subprocesses.
This allows to stop processing in case we failed to do the Change Data
Capture setup, which involves creating a replication slot with the logical
decoding plugin wal2json. When that plugin is not available on the source
database, it's best to exit early with a clear error message.

Without this change pgcopydb clone --follow would implement the full base
copy fo the database and only then exit with an error because the follow
subprocess failed early.