Implement same-table concurrency.
The idea is to split (or partition) larger tables into many COPY processes.
The threshold --split-tables-larger-than 100GB is also used to compute how
many processes to start for any single given table (table size / threshold).

When using same-table concurrency then we can't TRUNCATE in the same
transaction as where the COPY happens, because we now have several
concurrent connections doing the COPY. It means we can't use the COPY FREEZE
option anymore when splitting the COPY operation in concurrent processes.

The truncate operation still happens, it is implemented in a critical
section by the first concurrent COPY process that gets there.

The COPY source partitioning is computed with a SQL query that uses the SRF
generate_series, and then a COPY (SELECT ...) form is used to restrict the
source COPY to the computed range. Ranges are non-overlapping by construct
so we use (SELECT * FROM table WHERE partKey BETWEEN min AND max).

At the moment the partKey must be a single integer column (int2, int4, or
int8) with a unique constraint (either UNIQUE or PRIMARY KEY). Tables larger
than the threshold --split-tables-larger-than but without a candidate
partKey are still handled by a single process for the COPY operation.