Implement --debug and --trace, adding to --verbose. (#97)
Some users are confused by the Unix way of using -vvv to mean “very, very
verbose” and accumulate verbosity that way. Introduce --debug as a synonym
to -vv and --trace as a synonym to -vvv.

The only drawback to also implement --debug and --trace when we could use
--verbose several times is to use yet another short option letter. It seems
to be worth it given the user feedback though.

Also introduce a new log level NOTICE between INFO (default verbosity) and
DEBUG (extra verbosity) so that we have a new verbose setting that makes
sense. The log_notice() macro isn't used much as yet, following patches and
PRs will introduce more usage for it.