Add support for Extension Configuration Tables (#101)
* Implement pgcopydb list extensions.

* Implement pgcopydb copy extensions.

* Add support for extension configuration in pgcopydb copy extensions.

* Fix extension listing query.

* Add a unit test covering extensions, and fix some bugs.

* Fix schema queries for sequences when filtering is involved.

* Add schema/data that requires Postgis in the extensions unit test.

* Remove extra logging that's not necessary anymore.

* Implement pgcopydb clone --skip-extensions.

This allows to skip creating extensions on the target database, which
requires superuser privileges. This could now be implemented with another
command such as pgcopydb copy extensions.

* Fix connection management for pgcopydb copy extensions.

* Logs and comments clean-up.

* Make sure we can --skip-extensions and use a non-superuser role.

The test case now uses non-superuser for the pgcopydb clone operation, and
deals with roles and extensions separately, with superuser privileges. For
that to work, we need to skip restoring the ACL and COMMENT entries for the
extensions and the schema they depend on, as that's superuser territory.

* Copy extension configuration table user-data in pgcopydb clone.

* Add a documentation section about Postgres superuser.