Use docker-compose run in tests, instead of docker-compose up.
In passing, also change the test build dependencies in a way that
docker-compose knows about our usage of the common pagila container. This
allows skipping useless builds of that intermediate container.

The docker-compose run command also is a better UI for our testing when
compared to using docker-compose up and then waiting until the main
container stops.