Use docker-compose run in tests, instead of docker-compose up. (#113)
* Use docker-compose run in tests, instead of docker-compose up.

In passing, also change the test build dependencies in a way that
docker-compose knows about our usage of the common pagila container. This
allows skipping useless builds of that intermediate container.

The docker-compose run command also is a better UI for our testing when
compared to using docker-compose up and then waiting until the main
container stops.

* Separate our build pgcopydb service and the test service.

* Attempt to fix GitHub worklow for tests: compute version string first.

* Review and simplify docker-compose build dependencies.

In short docker-compose does not know how to handle build time dependencies
between different containers. I suppose we would need a place to push the
intermediate images and pull them later, and we're now using the local
docker daemon for that by building the pre-requisites manually before
invoking the docker-compose build and run commands.