Make sure to transform JSON files in all cases.
There is a case when we reach the endpos at startup (or restart), and we
have a JSON file around that was not transformed yet: the current LSN is not
past this file yet, and the previous time when streaming we didn't reach to
the endpos yet.

Now we need to transform that JSON file into SQL, even though we don't have
an open file pointer to call fclose() on.