Implement Logical Decoding compatibility with Postgres 9.6.
Several function and column names have changed (xlog to wal) between
Postgres major versions 9.6 and 10, and also in Postgres 9.6 creating a
logical replication slot seems to always export a (new) snapshot.

In this PR we adjust pgcopydb behavior to rely on the logical replication
protocol CREATE_REPLICATION_SLOT command to export our snapshot when the
option --follow is used and the source server is Postgres 9.6.