Implement an option to pg_table_size() results.
In some environments computing the pg_table_size() can be quite slow, so we
might benefit from managing a cache of the table sizes that we can re-use.
This cache is implemented as the pgcopydb.table_size table on the source
database, which is created with the command pgcopydb list tables --cache.