Fix the Dockerfile for wal2json support with older Postgres versions.
The apt.postgresql.org debian repository has archived its support for debian
stretch release, in a way that we need to now use
apt-archive.postgresql.org.

We could switch to using a bare debian system as the base for our docker
image here, but we would then have to provide and maintain the entry points
scripting that the official docker image is providing.