Arrange to use Logical Replication protocol metadata. (#155)
Instead of relying on wal2json to include the metadata we need in our
implementation of Change Data Capture, use the Logical Decoding protocol
metadata directly.

This prepares for compatibility with other output_plugins in the future.

The XID is still parsed from the "include-xids" option of wal2json at the
moment, but it's used only as a hint and a check. The current plan is to
have compatilibity with only output_plugins that can send the XIDs, such as
for example the test_decoding output_plugin.