Implement support for logical decoding plugin test_decoding. (#156)
By default Postgres ships with the test_decoding contrib module. While this contrib is meant only for testing the logical decoding APIs, it has been found useful to just rely on it in many situations. One key advantage compared to wal2json is that the data representation used in test_decoding is exactly the output from Postgres data type functions, without edits, so we know it will be accepted as-is by the target Postgres system.

This commit introduces the --plugin option, with support for either "wal2json" or "test_decoding". The default for the --plugin option is now "test_decoding".