Switch producers to line buffered output mode.
This avoids partial reads of a part of a line, now we read a line at a time,
unless it doesn't fit in our 128 MB read buffer, a case that is handled in
the code already (albeit quite poorly at the time being).

Also improve logging, install set -o pipefail in the test suite.