Implement new command: pgcopydb ping
The main use case is in the test-suite where we can use our retry policy to
make sure the environment is ready for running the test script. Because it's
useful for the project in the context of its unit testing suite, it might be
useful also for any project using pgcopydb with some automation around it.