Refactor the transform process management. (#194)
* Refactor the transform process management.

Make the transform process a sibling process to the receive and catchup
processes, where it was started by the receive process before. This
refactoring simplifies our process management and prepares for being able to
transition from the prefetch model to the live replay model.

* More refactoring, also name our queues.

Improve debugging and process management, and share more code to start the
follow sub-processes and supervise them, in a way that the same supervision
code applies to many of our modes of operations: prefetch, prefetch and
catchup, and soon, replay.