Implement pgcopydb stream replay.
The idea is to have an integrated command that does the same as the
following unix pipe command:

    $ pgcopydb stream receive --to-stdout
      | pgcopydb stream transform - -
	  | pgcopydb stream apply -

The command takes care of setting up the unix pipes and the process tree
internally. Makes it easier to use, and the implementation is needed for
intagration in the end-user API (pgcopydb clone --follow) anyway.