Refactor our internal representation for Logical Messages. (#198)
The KEEPALIVE and SWITCH WAL messages could be part of a
transaction (typically one that spans multiple WAL files), but also can be
found independently, between transactions.

Our previous internal representation insisted that every message be put in
the context of a transaction, but that doesn't fit well with the reality.

This refactoring is also necessary for streaming, where we don't want to
wait until the end of the next transaction before replaying the KEEPALIVE
messages. That kind of defeat the purpose of a KEEPALIVE message.