Review follow mode subprocess management.
Instead of sending a TERM signal to other sub-processes as soon as the first
one has finished, make sure that we detect feof() on the stream those
processes are reading from.

Also make sure to propagate the --endpos to the pgcopydb.sentinel table
before starting the sub-processes, so that they can all have the same
information to share. In passing, if the endpos has been reached already at
startup, skip starting the sub-processes entirely.