Fix tests.
The pgcopydb list table-parts code needed to know about the new facility
where we may create a temporary table and then need to prepend pgcopydb to
the search_path.

Also refactor the new logic in schema_prepare_pgcopydb_table_size() to make
it easier to follow. It was getting too complex to read for what it does.