Improve logs. (#215)
Add a new log level, SQL, that sits between NOTICE and DEBUG. In practice
the SQL queries and parameters are now logged when using --debug command
line, or when using --notice --verbose.

The option --notice is new and a synonym to --verbose. It has been added
because the --verbose option activates the NOTICE log level.

The logs time format was a short format including only the time, bypassing
the year/month/date parts because it takes a lot of terminal space and isn't
useful in an interactive context. This patch introduces support for the
environment variable PGCOPYDB_LOG_TIME_FORMAT which can host any format
string for the strftime(3). Also, the default log time format is
automatically switched to full date when logging to file (isatty is false).

Finally, when logging to file, then the FILE and LINE information are always
added to the output.