Ensure clean-up of System V resources.
Sytem V resources are created for sub-process communication within pgcopydb
run-time. At exit, we need to ensure that we clean-up those resources, which
are not automatically reclaimed by the OS.

We used to do that correctly with an atexit(3) handle function for the main
logging semaphore, now this is done systematically for all the semaphores
and queues created. This also ensures that the same process creates and
unlinks the resources, which wasn't always the case in the previous coding.