Implement a --fail-fast option.
This option allows pgcopydb to abort as soon as an error happens in one of
the sub-processes. For this, a TERM signal is sent to the process group. The
option can also be set from the environment variable PGCOPYDB_FAIL_FAST.

In passing, improve the sequence setting code for error management. As soon
as a sequence reset operation has failed, the Postgres transaction is
broken. Use savepoints to allow the error management code to continue then.