Fix how we skip concurrent build of certain indexes.
Even though the create index step must be skipped in some cases (when
Postgres doesn't have support for separate CREATE INDEX and ALTER TABLE
commands), we should still track the operation as if it was done, which
allows the code to later build the constraint with ALTER TABLE syntax.
That's when the index gets created actually.

This also allows to then skip the pg_restore entry for both the index and
the attached constraint.