Refrain from logging user data.
In the apply process logging the SQL statements opens the door to logging
user data. For privacy concerns, reduce the apply module logging.

Also reduce the amount of logs sent by the receive and apply processes in
general, as this can lead to a very large volume of data and we do not
anciticape needing that level of details to be able to debug operations.