Use dynamic memory for variable length schema parts. (#249)
Rather than using a static BUFSIZE string buffer we call calloc() to
allocate memory dynamically for index columns, index definitions, constraint
definition, and collation descriptions.

The allocated memory is used throughout the pgcopydb program execution, and
at the moment the memory is not de-allocated by a call to free(). Instead,
we rely on the OS to take care of that for us when exiting pgcopydb.