Transform empty transactions from the JSON to the SQL files. (#251)
Recent changes had the impact of discarding empty transactions entrirely
from the SQL file when transforming them from the JSON file. These
transactions act somewhat like a KEEPALIVE message of sorts, as they
represent activity in the source database that is not happening in the
database we're following.

In this commit we force the transform of these empty transactions to make
them part of the SQL file.